+++ src/Contracts/LocalizedRoute.php
@@ -2,12 +2,13 @@
 
 namespace Goodcat\L10n\Contracts;
 
+use Goodcat\L10n\Routing\RouteTranslations;
 use Illuminate\Routing\Route;
 
 interface LocalizedRoute
 {
     /** @see \Goodcat\L10n\Mixin\LocalizedRoute::lang */
-    public function lang(): string;
+    public function lang(): Route|RouteTranslations;
 
     /** @see \Goodcat\L10n\Mixin\LocalizedRoute::getLocalizedName */
     public function getLocalizedName(string $locale): ?string;
@@ -17,4 +18,7 @@ interface LocalizedRoute
      * @see \Goodcat\L10n\Mixin\LocalizedRoute::makeTranslations
      */
     public function makeTranslations(): array;
+
+    /** @see \Goodcat\L10n\Mixin\LocalizedRoute::uriWithoutPrefix */
+    public function uriWithoutPrefix(): string;
 }