+++ src/L10n.php
@@ -7,7 +7,6 @@ use Goodcat\L10n\Contracts\LocalizedRouter;
 use Goodcat\L10n\Resolvers\BrowserPreferredLocale;
 use Goodcat\L10n\Resolvers\PreferredLocaleResolver;
 use Goodcat\L10n\Resolvers\UserPreferredLocale;
-use Goodcat\L10n\Routing\RouteTranslations;
 use Illuminate\Routing\Route;
 use Illuminate\Routing\Router;
 
@@ -29,30 +28,13 @@ class L10n
 
         /** @var Route&LocalizedRoute $route */
         foreach ($collection->getRoutes() as $route) {
-            /** @var RouteTranslations $translations */
-            $translations = $route->lang();
+            $translations = $route->lang()->fillMissing($route);
 
             if ($translations->isEmpty()) {
                 continue;
             }
 
             $router->forget($route);
-
-            $key = 'routes.' . ltrim($route->uri(), "{$route->getPrefix()}/");
-
-            foreach ($translations->genericLocales() as $locale) {
-                if (trans()->hasForLocale($key, $locale)) {
-                    $translations->addTranslations([
-                        $locale => trans($key, locale: $locale),
-                    ]);
-                }
-            }
-
-            $translations->addTranslations([
-                app()->getFallbackLocale() => config('l10n.hide_default_locale')
-                    ? str_replace($route->getPrefix(), '', $route->uri())
-                    : null
-            ]);
 
             foreach ($route->makeTranslations() as $localizedRoute) {
                 $collection->add($localizedRoute);